+++ internal/coverage/testdata/golden/go/pkg/calc.go
@@ -1,0 +1,8 @@
+package calc
+
+func Add(a, b int) int {
+	if a == 0 {
+		return b
+	}
+	return a + b
+}